refactor: Bowling 삭제
